fix(worker): explicit integer cast on duration_ms parameter for supabase pooler

+++ src/worker/execution/execute-run.ts
@@ -33,8 +33,8 @@ async function writeStepRun(
         (run_id, workspace_id, node_id, node_type, status,
          input_json, output_json, error_json, duration_ms, attempt,
          started_at, finished_at)
-      values ($1, $2, $3, $4, $5, $6::jsonb, $7::jsonb, $8::jsonb, $9, $10,
-              now() - ($9 * interval '1 millisecond'), now())
+      values ($1, $2, $3, $4, $5, $6::jsonb, $7::jsonb, $8::jsonb, $9::integer, $10,
+              now() - ($9::integer * interval '1 millisecond'), now())
     `,
     [
       params.runId,
